requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
cryptography==42.0.8
cssselect2==0.7.0
dj-database-url==2.2.0
Django==5.0.6
django-classy-tags==4.1.0
django-cms==4.1.1
django-entangled==0.5.4
django-filebrowser==4.0.3
django-filer==3.1.3
django-formtools==2.5.1
django-fsm==2.8.2
django-grappelli==3.0.9
django-parler==2.3
django-polymorphic==3.1.0
django-richtextfield==1.6.2
django-sekizai==4.1.0
django-tinymce==4.0.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==3.0.0
djangocms-file==3.0.1
djangocms-frontend==1.3.2
djangocms-googlemap==2.2.0
djangocms-picture==4.1.1
djangocms-snippet==3.1.1
djangocms-style==3.1.0
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.2
djangocms-video==3.1.0
easy-thumbnails==2.8.5
filebrowser==1.1.4
Flask==3.0.3
gunicorn==22.0.0
html5lib==1.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
reportlab==4.2.0
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.1
webencodings==0.5.1
Werkzeug==3.0.3
whitenoise==6.6.0
